search cache prompt buy spinner show turn message route plan timeout enlarge route map add customize mark clean pending download request revise setStatus of DownloadRequest

+++ NaviUtil/LocationSimulator.h
@@ -28,7 +28,7 @@ typedef enum
 }
 @property (nonatomic, weak) id<CLLocationManagerDelegate> delegate;
 @property (nonatomic) NSTimeInterval locationUpdateInterval;
-@property (nonatomic) int simulationSpeed;
+@property (nonatomic) double simulationSpeed;
 
 @property (readonly) bool isStart;
 @property (nonatomic) LocationSimulator_Type type;
